// src/app/Modules/rrhh/model/ypuestrabv.dto.ts
// Generated by https://quicktype.io

export interface YPuesTrabVDTO {
    codAdminProc: string;
    codAgrupaPto: number;
    codEscalaPto: string;
    codRelaemple: string;
    codiagru:     number;
    codicate:     number;
    codiconv:     number;
    codicoti:     number;
    codiepig:     number;
    codigrup:     string;
    codimutu:     number;
    codindes:     number;
    codinesp:     number;
    codinive:     string;
    codiprop:     string;
    coditari:     number;
    codocupa:     string;
    descpues:     string;
    extingui:     string;
    fechmodi:     string;
    finivers:     string;
    horamodi:     string;
    id:           ID;
    singpues:     string;
    tipoadmin:    number;
    tipomodelo:   number;
}

export interface ID {
    codienti: number;
    codipues: number;
    coditrpt: string;
    verspues: number;
}