// src/app/Modules/rrhh/model/yfasepuesv.dto.ts
// Generated by https://quicktype.io

import { Ydotacionv } from "./ydotacionv.dto";
import { YFasePuesDTO } from "./yfasepues.dto";

export class YFasePuesVDTO { 
    yfasepue:      YFasePuesDTO;
    BEspdes:        string;
    BPresup:        string;
    BRequis:        string;
    BRetrpue:       string;
    actividad2:     string;
    codiadpu:       string;
    codiagru:       number;
    codicate:       number;
    codiconv:       number;
    codicoti:       number;
    codiepig:       number;
    codigrup:       string;
    codimutu:       number;
    codindes:       number;
    codinesp:       number;
    codinive:       string;
    codiprop:       string;
    codispue:       string;
    coditari:       number;
    codocupa:       string;
    compatib:       string;
    descpues:       string;
    estitula:       string;
    fechmodi:       Date | null;
    finivers:       Date | null;
    ffinvers:       Date | null;
    horamodi:       string;
    id:             YFasePuesVDTOID;
    jornredu:       string;
    porcdedi:       number;
    realizaLaboral: string;
    ydotacionv:     Ydotacionv;

    constructor() {
        this.yfasepue = new YFasePuesDTO();
        this.BEspdes=        '';
        this.BPresup=        '';
        this.BRequis=        '';
        this.BRetrpue=       '';
        this.actividad2=     '';
        this.codiadpu=       '';
        this.codiagru=       0;
        this.codicate=       0;
        this.codiconv=       0;
        this.codicoti=       0;
        this.codiepig=       0;
        this.codigrup=       '';
        this.codimutu=       0;
        this.codindes=       0;
        this.codinesp=       0;
        this.codinive=       '';
        this.codiprop=       '';
        this.codispue=       '';
        this.coditari=       0;
        this.codocupa=       '';
        this.compatib=       '';
        this.descpues=       '';
        this.estitula=       '';
        this.fechmodi=       null;
        this.finivers=       null;
        this.ffinvers=       null;
        this.horamodi=       '';
        this.id=             new YFasePuesVDTOID();
        this.jornredu=       '';
        this.porcdedi=       100;
        this.realizaLaboral= '';
        this.ydotacionv=     new Ydotacionv();        
    }

}

export class YFasePuesVDTOID {
    codienti: number;
    codipues: number;
    coditrpt: string;
    numedota: number;
    numefase: number;
    versfase: number;

    constructor() {
        this.codienti= 1;
        this.codipues= 0;
        this.coditrpt= '4';
        this.numedota= 0;
        this.numefase= 0;
        this.versfase= 0;
    }
}