// src/app/Modules/rrhh/model/srvprev.dto.ts
// Generated by https://quicktype.io

import { NaturelaDTO } from "./naturela.dto";
import { SituadminDTO } from "./situadmin.dto";

export interface SrvprevDTO {
    aniossrv:  number;
    certacti:  string;
    codicate:  number;
    codiconv:  number;
    codindes:  number;
    codinesp:  number;
    codinive:  string;
    codirela:  string;
    codisiad:  number;
    cueradmi:  string;
    descplaz:  string;
    diassrv:   number;
    ffinanti:  Date;
    finianti:  Date;
    id:        SrvprevIdDTO;
    mesessrv:  number;
    porcdedi:  number;
    restado:   string;
    sitadm:    SituadminDTO;
    trienios:  number;
    validado:  string;
    ynaturela: NaturelaDTO;
}

export interface SrvprevIdDTO {
    codiempl: number;
    codienti: number;
    codiserv: number;
}